refactor: AgenticLoop tool execution → ToolCallProcessor 클래스 추출
AgenticLoop에서 tool_use 블록 실행 로직(sequential/parallel dispatch,
tiered safety classification, batch cost approval, error recovery,
consecutive failure tracking, token guard)을 ToolCallProcessor로
분리하여 SRP 준수. AgenticLoop은 대화 루프에만 집중.

- ToolCallProcessor: tool_executor.py에 추가 (executor, op_logger,
  error_recovery, hooks, mcp_manager, transcript 의존성 주입)
- _guard_tool_result: agentic_loop.py → tool_executor.py 이동
- AgenticLoop: _tool_processor 인스턴스 생성 후 process() 위임
- 테스트 3078개 전체 통과, lint/mypy 0 errors
